+++ app/routes/components/Tooltip/changelog.codegen.ts
@@ -4,4 +4,6 @@
 export const changelog = ''
   + '- 2025-09-25: Initial version.\n'
   + '- 2025-09-30: Support `placement` prop for tooltip positioning.\n'
+  + '- 2025-10-10: Move styles into `components` CSS layer. Code formatting\n'
+  + '  improvement.\n'
   ;