+++ app/routes/components/Tooltip/demos/Demo01.tsx
@@ -1,16 +1,12 @@
 import { Button } from "src/components/Button";
 import { Tooltip } from "src/components/Tooltip";
 
-export function Demo() {
+export default function Demo() {
   return (
     <Tooltip
       label="This is a tooltip"
       triggerRender={({ setRef, onOpen, onClose }) => (
-        <Button
-          ref={setRef}
-          onMouseEnter={onOpen}
-          onMouseLeave={onClose}
-        >
+        <Button ref={setRef} onMouseEnter={onOpen} onMouseLeave={onClose}>
           Hover me
         </Button>
       )}