+++ app/routes/components/Tooltip/index.tsx
@@ -1,15 +1,12 @@
-import type { Route } from "./+types/index";
-
 import { ComponentTemplate } from "src/components/ComponentTemplate";
 import { Section } from "src/components/Section";
-
-import { sourceCode } from "./source_code.codegen";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01SourceCode from "./demos/Demo01.source_code.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-import { Demo } from "./Demo";
-
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "Tooltip | Copy UI" },
     { name: "description", content: "The Tooltip component from Copy UI." },
@@ -19,12 +16,11 @@ export function meta({}: Route.MetaArgs) {
 export default function TooltipPage() {
   return (
     <ComponentTemplate component="Tooltip">
-      <Section
-        title="Tooltip"
-        demoAndCode={[<Demo />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="Tooltip">
+        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }