+++ src/components/Tooltip/Tooltip.tsx
@@ -1,9 +1,9 @@
 import type { FC, ReactNode } from "react";
 
-import { Popover, type Placement } from "../Popover";
+import { type Placement, Popover } from "../Popover";
 import { useTheme } from "../ThemeProvider";
 
-import styles from './Tooltip.module.scss';
+import styles from "./Tooltip.module.scss";
 
 interface TooltipTriggerRenderProps {
   setRef: (el: Element | null) => void;
@@ -19,21 +19,21 @@ interface TooltipProps {
 }
 
 export const Tooltip: FC<TooltipProps> = (props) => {
-  const { label, placement = 'top', triggerRender } = props;
+  const { label, placement = "top", triggerRender } = props;
 
   const theme = useTheme();
 
   const computedStyle = {
-    '--tooltip-background-color': theme.colors.gray['900'],
-    '--tooltip-color': theme.colors.gray['100'],
-  }
+    "--tooltip-background-color": theme.colors.gray["900"],
+    "--tooltip-color": theme.colors.gray["100"],
+  };
 
   return (
     <Popover placement={placement}>
       <Popover.Trigger render={triggerRender} />
       <Popover.Portal
         onClickOutside={({ closePortal }) => closePortal()}
-        render={({ setRef, isOpen, floatingStyles }) => (
+        render={({ setRef, isOpen, floatingStyles }) =>
           isOpen && (
             <div
               ref={setRef}
@@ -43,7 +43,7 @@ export const Tooltip: FC<TooltipProps> = (props) => {
               {label}
             </div>
           )
-        )}
+        }
       />
     </Popover>
   );