+++ src/components/Tooltip/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-09-25: Initial version.
 - 2025-09-30: Support `placement` prop for tooltip positioning.
+- 2025-10-10: Move styles into `components` CSS layer. Code formatting
+  improvement.